Cập nhật hoàn chỉnh -1

+++ BarBeeOrder/BarBeeOrder/Areas/Admin/Controllers/AdminPagesController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BarBeeOrder.Models;
 using AspNetCoreHero.ToastNotification.Abstractions;
+using Microsoft.AspNetCore.Http;
 
 namespace BarBeeOrder.Areas.Admin.Controllers
 {
@@ -24,34 +25,119 @@ namespace BarBeeOrder.Areas.Admin.Controllers
         // GET: Admin/AdminPages
         public async Task<IActionResult> Index()
         {
-            
-            var models = _context.Pages.OrderByDescending(x=> x.PageId).ToList();
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
+            {
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        var models = _context.Pages.OrderByDescending(x => x.PageId).ToList();
 
-            return View(models);
+                        return View(models);
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                   
+
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+
+
+            
         }
 
         // GET: Admin/AdminPages/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
             {
-                return NotFound();
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        if (id == null)
+                        {
+                            return NotFound();
+                        }
+
+                        var page = await _context.Pages
+                            .FirstOrDefaultAsync(m => m.PageId == id);
+                        if (page == null)
+                        {
+                            return NotFound();
+                        }
+
+                        return View(page);
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                    
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
             }
 
-            var page = await _context.Pages
-                .FirstOrDefaultAsync(m => m.PageId == id);
-            if (page == null)
-            {
-                return NotFound();
-            }
 
-            return View(page);
+            
         }
 
         // GET: Admin/AdminPages/Create
         public IActionResult Create()
         {
-            return View();
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
+            {
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        return View();
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+
+                    
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            
         }
 
         // POST: Admin/AdminPages/Create
@@ -61,32 +147,86 @@ namespace BarBeeOrder.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PageId,PageName,PageContent,Published,Tittle,Ordering,CreatedDate,IsHeader")] Page page)
         {
-            if (ModelState.IsValid)
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
             {
-                page.CreatedDate = DateTime.Now;    
-                
-                _context.Add(page);
-                await _context.SaveChangesAsync();
-                _notyfService.Success("Tạo mới thành công!");
-                return RedirectToAction(nameof(Index));
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        if (ModelState.IsValid)
+                        {
+                            page.CreatedDate = DateTime.Now;
+
+                            _context.Add(page);
+                            await _context.SaveChangesAsync();
+                            _notyfService.Success("Tạo mới thành công!");
+                            return RedirectToAction(nameof(Index));
+                        }
+                        return View(page);
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                    
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
             }
-            return View(page);
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            
         }
 
         // GET: Admin/AdminPages/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
             {
-                return NotFound();
-            }
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        if (id == null)
+                        {
+                            return NotFound();
+                        }
 
-            var page = await _context.Pages.FindAsync(id);
-            if (page == null)
-            {
-                return NotFound();
+                        var page = await _context.Pages.FindAsync(id);
+                        if (page == null)
+                        {
+                            return NotFound();
+                        }
+                        return View(page);
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                    
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
             }
-            return View(page);
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            
         }
 
         // POST: Admin/AdminPages/Edit/5
@@ -96,51 +236,105 @@ namespace BarBeeOrder.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PageId,PageName,PageContent,Published,Tittle,Ordering,CreatedDate,IsHeader")] Page page)
         {
-            if (id != page.PageId)
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
             {
-                return NotFound();
-            }
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        if (id != page.PageId)
+                        {
+                            return NotFound();
+                        }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(page);
-                    _notyfService.Success("Chỉnh sửa thành công!");
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PageExists(page.PageId))
-                    {
-                        return NotFound();
+                        if (ModelState.IsValid)
+                        {
+                            try
+                            {
+                                _context.Update(page);
+                                _notyfService.Success("Chỉnh sửa thành công!");
+                                await _context.SaveChangesAsync();
+                            }
+                            catch (DbUpdateConcurrencyException)
+                            {
+                                if (!PageExists(page.PageId))
+                                {
+                                    return NotFound();
+                                }
+                                else
+                                {
+                                    throw;
+                                }
+                            }
+                            return RedirectToAction(nameof(Index));
+                        }
+                        return View(page);
                     }
-                    else
+                    catch
                     {
-                        throw;
+                        return RedirectToAction("Error", "Error", new { area = "" });
                     }
+                    
+
                 }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", "Home", new { area = "" });
             }
-            return View(page);
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            
         }
 
         // GET: Admin/AdminPages/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
             {
-                return NotFound();
-            }
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        if (id == null)
+                        {
+                            return NotFound();
+                        }
 
-            var page = await _context.Pages
-                .FirstOrDefaultAsync(m => m.PageId == id);
-            if (page == null)
+                        var page = await _context.Pages
+                            .FirstOrDefaultAsync(m => m.PageId == id);
+                        if (page == null)
+                        {
+                            return NotFound();
+                        }
+
+                        return View(page);
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                   
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            else
             {
-                return NotFound();
+                return RedirectToAction("Index", "Home", new { area = "" });
             }
-
-            return View(page);
+            
         }
 
         // POST: Admin/AdminPages/Delete/5
@@ -148,11 +342,39 @@ namespace BarBeeOrder.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var page = await _context.Pages.FindAsync(id);
-            _context.Pages.Remove(page);
-            await _context.SaveChangesAsync();
-            _notyfService.Warning("Xóa thành công!");
-            return RedirectToAction(nameof(Index));
+            var taikhoanID = HttpContext.Session.GetString("CustomerId");
+            if (taikhoanID != null)
+            {
+                var khachhang = _context.Customers.AsNoTracking().SingleOrDefault(x => x.CustomerId == Convert.ToInt32(taikhoanID));
+                if (khachhang != null)
+                {
+                    if (khachhang.RoleId == 2)
+                    {
+                        return RedirectToAction("Index", "Home", new { area = "" });
+                    }
+                    try
+                    {
+                        var page = await _context.Pages.FindAsync(id);
+                        _context.Pages.Remove(page);
+                        await _context.SaveChangesAsync();
+                        _notyfService.Warning("Xóa thành công!");
+                        return RedirectToAction(nameof(Index));
+                    }
+                    catch
+                    {
+                        return RedirectToAction("Error", "Error", new { area = "" });
+                    }
+                   
+
+
+                }
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home", new { area = "" });
+            }
+            
         }
 
         private bool PageExists(int id)
